+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * potentially long running Git actions are now made as a scheduled task no longer limited by web request time
 * configurable Git repository path
+* view template's content endpoints
   
 #### Improvements